Fix handling of dependencies without any licenses at all (#44)
We made an incorrect assumption that "none" would be substituted in by
Composer (as the text output of `composer licenses` does) when this
situation happens.

Fixes 2 issues:

- `composer licenses` returns an empty list of licenses when a
dependency does not specify a license at all. Our parsing here would
result in always accepting that dependency (as we were iterating the
licenses, so had nothing to reject).
- `installed.json` does not contain the `license` key when a dependency
does not specify a license at all. Our parsing would error out entirely
on missing key when this was encountered.

This fixes both issues, rejecting any depedencies which do not specify a
license, and are not explicitly allowed as a vendor or package in
configuration.

---------

Co-authored-by: Marcin Michalski <marcin+github@michalski.dev>